refactor(game): move result and score calculators to model package
