+++ src/main/kotlin/model/ScoreCalculator.kt
@@ -1,7 +1,4 @@
-package service
-
-import model.Card
-import model.Rank
+package model
 
 class ScoreCalculator {
     companion object {
